# api/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
    "/api/admin/users" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserAdminView"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List",
        "tags" : [ "User Admin Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN" ]
        } ]
      }
    },
...
    "/api/me" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CurrentUserView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Current User",
        "tags" : [ "User Controller" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
    "/api/me/profile" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Profile",
        "tags" : [ "Profile Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
    "/api/me/profile/notification-preferences" : {
      "put" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NotificationPreferences"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update Notification Preferences",
        "tags" : [ "Profile Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
    "/api/me/profile/office" : {
      "put" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OfficePreference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OfficePreference"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update Office",
        "tags" : [ "Profile Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
    "/api/users" : {
      "get" : {
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "maxLength" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserSummary"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List Users",
        "tags" : [ "User Controller" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
...
    "/talks" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Talk"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List",
        "tags" : [ "Talk Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Talk"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Talk"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create",
        "tags" : [ "Talk Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
    "/talks/review" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TalkReviewRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TalkReviewResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Review",
        "tags" : [ "Talk Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      }
    },
    "/talks/{id}" : {
      "put" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Talk"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Talk"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update",
        "tags" : [ "Talk Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Talk"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get",
        "tags" : [ "Talk Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN", "DT", "CONSULTANT" ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Delete",
        "tags" : [ "Talk Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "ADMIN" ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Conference" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "date" : {
            "$ref" : "#/components/schemas/ConferencePeriod"
          },
          "cfpLink" : {
            "type" : "string"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "cfpStatus" : {
            "type" : "string"
          },
          "submittedTalksAmount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "cfpClosingDate" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/ConferenceType"
          },
          "reach" : {
            "$ref" : "#/components/schemas/ConferenceReach"
          }
        },
        "required" : [ "name" ]
      },
      "ConferencePeriod" : {
        "type" : "object",
        "required" : [ "start", "end", "precision" ],
        "properties" : {
          "start" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "end" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "precision" : {
            "$ref" : "#/components/schemas/DatePrecision"
          }
        }
      },
      "ConferenceReach" : {
        "type" : "string",
        "enum" : [ "Locale", "Régionale", "Nationale" ]
      },
      "ConferenceType" : {
        "type" : "string",
        "enum" : [ "Marketing / business", "Technique stratégique", "Technique généraliste", "Technique", "Hors scope" ]
      },
      "CurrentUserView" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "roles" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "DatePrecision" : {
        "type" : "string",
        "enum" : [ "DAY", "MONTH" ]
      },
...
      "Location" : {
        "type" : "object",
        "properties" : {
          "city" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "postalCode" : {
            "type" : "string"
          }
        }
      },
...
      "NotificationPreferences" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "boolean"
          },
          "slack" : {
            "type" : "boolean"
          }
        }
      },
      "Office" : {
        "type" : "string",
        "enum" : [ "paris", "nantes", "rennes", "bordeaux", "lyon", "lille", "grenoble", "singapour", "montreal" ]
      },
      "OfficePreference" : {
        "type" : "object",
        "properties" : {
          "office" : {
            "$ref" : "#/components/schemas/Office"
          }
        }
      },
      "ProfileView" : {
        "type" : "object",
        "required" : [ "name", "email", "notificationPreferences" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "email" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "notificationPreferences" : {
            "$ref" : "#/components/schemas/NotificationPreferences"
          },
          "slackLinked" : {
            "type" : "boolean"
          },
          "office" : {
            "$ref" : "#/components/schemas/Office"
          }
        }
      },
      "Role" : {
        "type" : "string",
        "enum" : [ "ADMIN", "DT", "CONSULTANT" ]
      },
...
      "Talk" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "description" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "speakers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/User"
            },
            "minItems" : 1
          },
          "office" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "conference" : {
            "$ref" : "#/components/schemas/Conference"
          },
          "status" : {
            "$ref" : "#/components/schemas/TalkStatus"
          },
          "visibility" : {
            "$ref" : "#/components/schemas/Visibility"
          },
          "format" : {
            "type" : "string",
            "anyOf" : [ {
              "$ref" : "#/components/schemas/TalkFormat"
            }, {
              "type" : "string"
            } ],
            "description" : "Format proposé par TalkFormat, ou ancienne valeur conservée sans modification",
            "pattern" : "\\S"
          },
          "date" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "slides" : {
            "type" : "string",
            "pattern" : "^$|https?://[^\\s]+"
          },
          "replay" : {
            "type" : "string",
            "pattern" : "^$|https?://[^\\s]+"
          },
          "audience" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          }
        },
        "required" : [ "title", "description", "speakers", "office", "status", "visibility", "format" ]
      },
      "TalkFormat" : {
        "type" : "string",
        "enum" : [ "quicky", "conference", "workshop" ]
      },
      "TalkReviewRequest" : {
        "type" : "object",
        "required" : [ "title", "abstract" ],
        "properties" : {
          "title" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "abstract" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }
      },
      "TalkReviewResponse" : {
        "type" : "object",
        "required" : [ "suggestedTitles", "suggestedAbstracts", "feedback", "keyImprovements" ],
        "properties" : {
          "suggestedTitles" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "suggestedAbstracts" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "feedback" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keyImprovements" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "TalkStatus" : {
        "type" : "string",
        "enum" : [ "DRAFT", "PLANNED", "SUBMITTED", "ACCEPTED", "REJECTED", "DONE" ]
      },
      "User" : {
        "type" : "object",
        "required" : [ "name" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "email" : {
            "type" : "string"
          },
          "slackUserId" : {
            "type" : "string"
          }
        }
      },
      "UserAdminView" : {
        "type" : "object",
        "required" : [ "name", "email", "roles" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "email" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Role"
            }
          }
        }
      },
      "UserSummary" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        }
      },
      "Visibility" : {
        "type" : "string",
        "enum" : [ "PUBLIC", "PRIVATE" ]
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Authentication"
      }
    }
  },
  "info" : {
    "title" : "Thezaurus API",
    "version" : "1.0.0-SNAPSHOT"
  }
}